engine/pythonProject/requirements.txt
alphashape==1.3.1
gensim==4.3.1
matplotlib==3.6.3
meshio==5.3.4
MeshPy==2022.1.3
numpy==1.23.5
numpy-stl==3.0.0
opencv-contrib-python==4.7.0.68
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
plyfile==0.7.4
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
tensorflow==2.15.0
trimesh==3.18.1
tripy==1.0.0
visvis==1.14.0
pydicom==2.3.1
pyspellchecker==0.7.1